@@ -9,7 +9,9 @@
         }
     ],
     "require": {
-        "php": ">=7.3"
+        "php": ">=8.1",
+        "ext-pdo": "*",
+        "ext-json": "*"
     },
     "autoload": {
         "psr-4": {
